add stubs for transformer rank + strategy
